requirements.txt
flask[async]
aiortc
pillow
fastapi
uvicorn
numpy
opencv-python
